+++ src/main/java/org/embulk/util/rubytime/RubyTimeFormat.java
@@ -1,4 +1,4 @@
-package org.embulk.spi.time;
+package org.embulk.util.rubytime;
 
 import java.util.ArrayList;
 import java.util.Collections;
