+++ src/main/java/org/embulk/util/rubytime/TimeZoneIds.java
@@ -1,4 +1,4 @@
-package org.embulk.spi.time;
+package org.embulk.util.rubytime;
 
 import java.time.DateTimeException;
 import java.time.ZoneId;
