+++ src/test/java/org/embulk/util/rubytime/TestRubyTimeParser.java
@@ -1,4 +1,4 @@
-package org.embulk.spi.time;
+package org.embulk.util.rubytime;
 
 import static org.junit.Assert.assertTrue;
